+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 + Ability to use utility from scripts
 
 ### Changed
++ Get information about artboards from resources directory
 + Separate functionality from CLI
 
 ## [Released]